Some refactoring of classes. Removing relationship
