+++ src/main/java/org/jdbctemplatemapper/annotation/Id.java
@@ -1,0 +1,12 @@
+package org.jdbctemplatemapper.annotation;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.FIELD)
+public @interface Id {
+    IdType type() default IdType.MANUAL;
+}
