+++ src/main/java/org/jdbctemplatemapper/core/JdbcTemplateMapper.java
@@ -1,17 +1,8 @@
 package org.jdbctemplatemapper.core;
 
-import java.beans.PropertyDescriptor;
-import java.lang.reflect.Field;
-import java.lang.reflect.Method;
-import java.lang.reflect.ParameterizedType;
-import java.lang.reflect.Type;
-import java.sql.DatabaseMetaData;
 import java.sql.ResultSet;
-import java.sql.ResultSetMetaData;
-import java.sql.SQLException;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
-import java.util.Collection;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.LinkedHashMap;
@@ -20,13 +11,9 @@ import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.atomic.AtomicInteger;
-import java.util.regex.Pattern;
-import java.util.stream.Collectors;
 
 import org.springframework.beans.BeanWrapper;
 import org.springframework.beans.PropertyAccessorFactory;
-import org.springframework.core.annotation.AnnotationUtils;
 import org.springframework.core.convert.support.DefaultConversionService;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.BeanPropertyRowMapper;
@@ -35,12 +22,8 @@ import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
-import org.springframework.jdbc.support.DatabaseMetaDataCallback;
 import org.springframework.jdbc.support.JdbcUtils;
-import org.springframework.jdbc.support.MetaDataAccessException;
 import org.springframework.util.Assert;
-import org.springframework.util.ReflectionUtils;
-import org.springframework.util.StringUtils;
 
 /**
  * <pre>
@@ -77,7 +60,7 @@ import org.springframework.util.StringUtils;
  *
  * // Product class below maps to product/PRODUCT table by default.
  * // Use annotation @Table(name="some_other_tablename") to override the default
- * 
+ *
  * public class Product {
  *    private Integer id; // 'id' property is needed for all models and has to be of type Integer or Long
  *    private String productName;
@@ -133,11 +116,11 @@ import org.springframework.util.StringUtils;
  *
  * &#64;Bean
  * public JdbcTemplateMapper jdbcTemplateMapper(JdbcTemplate jdbcTemplate) {
- * 
+ *
  *   return new JdbcTemplateMapper(jdbcTemplate);
- *   
+ *
  *   //if the database setup needs a schema name, pass it as argument.
- *   //return new JdbcTemplateMapper(jdbcTemplate, "your_database_schema_name");   
+ *   //return new JdbcTemplateMapper(jdbcTemplate, "your_database_schema_name");
  * }
  *
  * </pre>
@@ -149,27 +132,19 @@ public class JdbcTemplateMapper {
   private NamedParameterJdbcTemplate npJdbcTemplate;
   private IRecordOperatorResolver recordOperatorResolver;
 
-  private String catalogName;
-  private String schemaName;
+  // TONY private String catalogName;
+  // TONY private String schemaName;
   // this is for the call to databaseMetaData.getColumns() just in case a database needs something
   // other than null
-  private String metaDataColumnNamePattern;
+  // TONY private String metaDataColumnNamePattern;
   private String createdByPropertyName;
   private String createdOnPropertyName;
   private String updatedByPropertyName;
   private String updatedOnPropertyName;
   private String versionPropertyName;
-  
-  // to avoid query being issued with large number of ids
-  // for the sql 'IN' clause the id list is chunked by this size
-  // and multiple queries issued if needed.
-  private int inClauseChunkSize = 100;
 
   // Need this for type conversions like java.sql.Timestamp to java.time.LocalDateTime etc
   private DefaultConversionService defaultConversionService = new DefaultConversionService();
-
-  // Convert camel case to snake case regex pattern. Pattern is thread safe
-  private static Pattern TO_SNAKE_CASE_PATTERN = Pattern.compile("(.)(\\p{Upper})");
 
   // Inserts use SimpleJdbcInsert. Since SimpleJdbcInsert is thread safe, cache it
   // Map key - table name,
@@ -181,44 +156,22 @@ public class JdbcTemplateMapper {
   //     value - the update sql
   private Map<String, UpdateSqlAndParams> updateSqlAndParamsCache = new ConcurrentHashMap<>();
 
-  // Map key - table name,
-  //     value - the list of database column names
-  private Map<String, List<ColumnInfo>> tableColumnInfoCache = new ConcurrentHashMap<>();
-
-  // Map key - simple Class name
-  //     value - list of property names
-  private Map<String, List<PropertyInfo>> objectPropertyInfoCache = new ConcurrentHashMap<>();
-
-  // Map key - camel case string,
-  //     value - snake case string
-  private Map<String, String> camelToSnakeCache = new ConcurrentHashMap<>();
-
   // Map key - tableName-tableAlias
   //     value - the selectColumns string
   private Map<String, String> selectColumnsCache = new ConcurrentHashMap<>();
 
-  // Map key - object class name
-  //     value - the table name
-  private Map<String, TableMapping> objectToTableMappingCache = new ConcurrentHashMap<>();
+  private DatabaseUtils dbUtils;
+  private MappingUtils mappingUtils;
 
-  // Map key - object class name - propertyName
-  //     value - The getter Method
-  private Map<String, Method> objectGetMethodCache = new ConcurrentHashMap<>();
-
-  // Map key - object class name - propertyName
-  //     value - The setter Method
-  private Map<String, Method> objectSetMethodCache = new ConcurrentHashMap<>();
-
-  
   /**
    * The constructor.
    *
    * @param dataSource The dataSource for the mapper
    */
   public JdbcTemplateMapper(JdbcTemplate jdbcTemplate) {
-	  this(jdbcTemplate, null);
+    this(jdbcTemplate, null);
   }
-  
+
   /**
    * The constructor.
    *
@@ -228,8 +181,11 @@ public class JdbcTemplateMapper {
   public JdbcTemplateMapper(JdbcTemplate jdbcTemplate, String schemaName) {
     Assert.notNull(jdbcTemplate, "jdbcTemplate must not be null");
     this.jdbcTemplate = jdbcTemplate;
-    this.schemaName = schemaName;
     this.npJdbcTemplate = new NamedParameterJdbcTemplate(jdbcTemplate);
+
+    this.dbUtils = new DatabaseUtils(jdbcTemplate, npJdbcTemplate, schemaName);
+
+    this.mappingUtils = new MappingUtils(dbUtils);
   }
 
   /**
@@ -334,7 +290,7 @@ public class JdbcTemplateMapper {
    * @param catalogName The catalog
    */
   public void setCatalogName(String catalogName) {
-    this.catalogName = catalogName;
+    dbUtils.setCatalogName(catalogName);
   }
 
   /**
@@ -345,18 +301,7 @@ public class JdbcTemplateMapper {
    * @param metaDataColumnNamePattern
    */
   public void setMetaDataColumnNamePattern(String metaDataColumnNamePattern) {
-    this.metaDataColumnNamePattern = metaDataColumnNamePattern;
-  }
-  
-  /**
-   * Set the chunk size of the sql IN clause. Default is 100
-   * @param chunkSize
-   */
-  public void setInClauseChunkSize(Integer chunkSize) {
-	  if (chunkSize == null || chunkSize <= 0) {
-		  throw new IllegalArgumentException("chunkSize cannot be null and has to be greater than 0");
-	  }
-	  this.inClauseChunkSize = chunkSize;
+    dbUtils.setMetaDataColumnNamePattern(metaDataColumnNamePattern);
   }
 
   /**
@@ -373,7 +318,7 @@ public class JdbcTemplateMapper {
       throw new RuntimeException("id has to be type of Integer or Long");
     }
 
-    TableMapping tableMapping = getTableMapping(clazz);
+    TableMapping tableMapping = mappingUtils.getTableMapping(clazz);
     String idColumnName = tableMapping.getIdColumnName();
     if (idColumnName == null) {
       throw new RuntimeException(
@@ -385,7 +330,7 @@ public class JdbcTemplateMapper {
     }
     String sql =
         "SELECT * FROM "
-            + fullyQualifiedTableName(tableMapping.getTableName())
+            + dbUtils.fullyQualifiedTableName(tableMapping.getTableName())
             + " WHERE "
             + idColumnName
             + " = ?";
@@ -408,8 +353,8 @@ public class JdbcTemplateMapper {
   public <T> List<T> findAll(Class<T> clazz) {
     Assert.notNull(clazz, "Class must not be null");
 
-    String tableName = getTableMapping(clazz).getTableName();
-    String sql = "SELECT * FROM " + fullyQualifiedTableName(tableName);
+    String tableName = mappingUtils.getTableMapping(clazz).getTableName();
+    String sql = "SELECT * FROM " + dbUtils.fullyQualifiedTableName(tableName);
     RowMapper<T> mapper = BeanPropertyRowMapper.newInstance(clazz);
     return jdbcTemplate.query(sql, mapper);
   }
@@ -425,14 +370,16 @@ public class JdbcTemplateMapper {
   public <T> List<T> findAll(Class<T> clazz, String orderByClause) {
     Assert.notNull(clazz, "Class must not be null");
 
-    String tableName = getTableMapping(clazz).getTableName();
-    String sql = "SELECT * FROM " + fullyQualifiedTableName(tableName) + " " + orderByClause;
+    String tableName = mappingUtils.getTableMapping(clazz).getTableName();
+    String sql =
+        "SELECT * FROM " + dbUtils.fullyQualifiedTableName(tableName) + " " + orderByClause;
     RowMapper<T> mapper = BeanPropertyRowMapper.newInstance(clazz);
     return jdbcTemplate.query(sql, mapper);
   }
 
   /**
-   * Inserts an object whose id in database is auto increment. Once inserted the object will have the id assigned.
+   * Inserts an object whose id in database is auto increment. Once inserted the object will have
+   * the id assigned.
    *
    * <p>Also assigns created by, created on, updated by, updated on, version if these properties
    * exist for the object and the JdbcTemplateMapper is configured for them.
@@ -454,7 +401,7 @@ public class JdbcTemplateMapper {
           "For method insert() the objects 'id' property has to be null since this insert is for an object whose id is autoincrement in database.");
     }
 
-    TableMapping tableMapping = getTableMapping(obj.getClass());
+    TableMapping tableMapping = mappingUtils.getTableMapping(obj.getClass());
     if (tableMapping.getIdColumnName() == null) {
       throw new RuntimeException(
           "table " + tableMapping.getTableName() + " does not have a property named 'id'");
@@ -484,14 +431,14 @@ public class JdbcTemplateMapper {
       bw.setPropertyValue(versionPropertyName, 1);
     }
 
-    Map<String, Object> attributes = convertToSnakeCaseAttributes(obj);
+    Map<String, Object> attributes = CommonUtils.convertToSnakeCaseAttributes(obj);
 
     SimpleJdbcInsert jdbcInsert = simpleJdbcInsertCache.get(tableMapping.getTableName());
     if (jdbcInsert == null) {
       jdbcInsert =
           new SimpleJdbcInsert(jdbcTemplate)
-              .withCatalogName(catalogName)
-              .withSchemaName(schemaName)
+              .withCatalogName(dbUtils.getCatalogName())
+              .withSchemaName(dbUtils.getSchemaName())
               .withTableName(tableMapping.getTableName())
               .usingGeneratedKeyColumns(tableMapping.getIdColumnName());
       simpleJdbcInsertCache.put(tableMapping.getTableName(), jdbcInsert);
@@ -525,7 +472,7 @@ public class JdbcTemplateMapper {
           "For method insertById() the objects 'id' property cannot be null.");
     }
 
-    TableMapping tableMapping = getTableMapping(obj.getClass());
+    TableMapping tableMapping = mappingUtils.getTableMapping(obj.getClass());
     if (tableMapping.getIdColumnName() == null) {
       throw new RuntimeException(
           "table " + tableMapping.getTableName() + " does not have a property named 'id'");
@@ -558,14 +505,14 @@ public class JdbcTemplateMapper {
     if (jdbcInsert == null) {
       jdbcInsert =
           new SimpleJdbcInsert(jdbcTemplate)
-              .withCatalogName(catalogName)
-              .withSchemaName(schemaName)
+              .withCatalogName(dbUtils.getCatalogName())
+              .withSchemaName(dbUtils.getSchemaName())
               .withTableName(tableMapping.getTableName());
 
       simpleJdbcInsertCache.put(tableMapping.getTableName(), jdbcInsert);
     }
 
-    Map<String, Object> attributes = convertToSnakeCaseAttributes(obj);
+    Map<String, Object> attributes = CommonUtils.convertToSnakeCaseAttributes(obj);
     jdbcInsert.execute(attributes);
   }
 
@@ -585,7 +532,7 @@ public class JdbcTemplateMapper {
           "Object " + obj.getClass().getSimpleName() + " does not have a property named 'id'");
     }
 
-    TableMapping tableMapping = getTableMapping(obj.getClass());
+    TableMapping tableMapping = mappingUtils.getTableMapping(obj.getClass());
     String idColumnName = tableMapping.getIdColumnName();
     if (idColumnName == null) {
       throw new RuntimeException(
@@ -607,7 +554,7 @@ public class JdbcTemplateMapper {
 
       Set<String> updatePropertyNames = new LinkedHashSet<>();
 
-      for (PropertyInfo propertyInfo : getObjectPropertyInfo(obj)) {
+      for (PropertyInfo propertyInfo : CommonUtils.getObjectPropertyInfo(obj)) {
         // if not a ignore property and has a table column mapping add it to the update property
         // list
         if (!ignoreAttrs.contains(propertyInfo.getPropertyName())
@@ -639,7 +586,7 @@ public class JdbcTemplateMapper {
           "Object " + obj.getClass().getSimpleName() + " does not have a property named 'id'");
     }
 
-    TableMapping tableMapping = getTableMapping(obj.getClass());
+    TableMapping tableMapping = mappingUtils.getTableMapping(obj.getClass());
     String idColumnName = tableMapping.getIdColumnName();
     if (idColumnName == null) {
       throw new RuntimeException(
@@ -729,14 +676,16 @@ public class JdbcTemplateMapper {
           "Object " + obj.getClass().getSimpleName() + " has to have a property named 'id'.");
     }
 
-    TableMapping tableMapping = getTableMapping(obj.getClass());
+    TableMapping tableMapping = mappingUtils.getTableMapping(obj.getClass());
     if (tableMapping.getIdColumnName() == null) {
       throw new RuntimeException(
           "table " + tableMapping.getTableName() + " does not have a property named 'id'");
     }
 
     String sql =
-        "delete from " + fullyQualifiedTableName(tableMapping.getTableName()) + " where id = ?";
+        "delete from "
+            + dbUtils.fullyQualifiedTableName(tableMapping.getTableName())
+            + " where id = ?";
     Object id = bw.getPropertyValue("id");
     return jdbcTemplate.update(sql, id);
   }
@@ -754,757 +703,11 @@ public class JdbcTemplateMapper {
     if (!(id instanceof Integer || id instanceof Long)) {
       throw new RuntimeException("id has to be type of Integer or Long");
     }
-    String tableName = getTableMapping(clazz).getTableName();
-    String sql = "delete from " + fullyQualifiedTableName(tableName) + " where id = ?";
+    String tableName = mappingUtils.getTableMapping(clazz).getTableName();
+    String sql = "delete from " + dbUtils.fullyQualifiedTableName(tableName) + " where id = ?";
     return jdbcTemplate.update(sql, id);
   }
 
-  /**
-   * Populates the toOne relationship of the main object. Issues an sql query to get the
-   * relationship. Make sure the join property of the argument main object is assigned so it can be
-   * matched to its corresponding relationship object (mainObj.mainObjJoinPropertyName =
-   * relationshipObj.id)
-   *
-   * <pre>
-   * Main Object:
-   * public class Order{
-   *   Integer id;
-   *   LocalDateTime orderDate;
-   *   Integer customerId;
-   *   Customer customer; // the toOne relationship
-   * }
-   * toOne related Object:
-   * public class Customer{
-   *   Integer id;
-   *   String firstName;
-   *   String lastName;
-   *   String address;
-   * }
-   *
-   * 1) Get an Order Object using a query for example:
-   * Order order = jdbcTemplateMapper.findById(orderId, Order.class);
-   *
-   * 2) Populate the Order's toOne customer property. This will issue an sql and populate order.customer
-   * jdbcTemplateMapper.toOne(order,"customer", "customerId");
-   *
-   * </pre>
-   *
-   * @param mainObj the main object
-   * @param mainObjRelationshipPropertyName The propertyName of the toOne relationship (on mainOjb)
-   *     that needs to be populated.
-   * @param mainObjJoinPropertyName the join property on main object.
-   */
-  public <T> void toOne(
-      T mainObj, String mainObjRelationshipPropertyName, String mainObjJoinPropertyName) {
-
-    Assert.notNull(
-        mainObjRelationshipPropertyName, "mainObjRelationshipPropertyName must not be null");
-    Assert.notNull(mainObjJoinPropertyName, "mainObjJoinPropertyName must not be null");
-
-    if (mainObj != null) {
-      List<T> mainObjList = new ArrayList<>();
-      mainObjList.add(mainObj);
-      toOne(mainObjList, mainObjRelationshipPropertyName, mainObjJoinPropertyName);
-    }
-  }
-
-  /**
-   * Populates the toOne relationship for all the main objects in the argument list. Issues an sql
-   * query using the 'IN' clause to get all the relationship objects corresponding to the main
-   * object list. Make sure the join property of the main object is assigned so it can be matched to
-   * its corresponding relationship object (mainObj.mainObjJoinPropertyName = relationshipObj.id)
-   *
-   * <pre>
-   * Main Object:
-   * public class Order{
-   *   Integer id;
-   *   LocalDateTime orderDate;
-   *   Integer customerId;
-   *   Customer customer; // the toOne relationship
-   * }
-   * toOne related Object:
-   * public class Customer{
-   *   Integer id;
-   *   String firstName;
-   *   String lastName;
-   *   String address;
-   * }
-   *
-   * 1) Get list of orders using a query like below or a complex query using jdbcTemplate.
-   * List<Order> orders = jdbcTemplateMapper.findAll(Order.class)
-   *
-   * 2) Populate each Order's toOne customer property. This will issue an sql to get the corresponding
-   *    customers using an IN clause.
-   * jdbcTemplateMapper.toOne(orders, "customer", "customerId");
-   *
-   * </pre>
-   *
-   * @param mainObjList list of main objects
-   * @param mainObjRelationshipPropertyName The toOne relationship property name on main object
-   * @param mainObjJoinPropertyName the join property name on the main object.
-   */
-  @SuppressWarnings("all")
-  public <T> void toOne(
-      List<T> mainObjList, String mainObjRelationshipPropertyName, String mainObjJoinPropertyName) {
-
-    Assert.notNull(
-        mainObjRelationshipPropertyName, "mainObjRelationshipPropertyName must not be null");
-    Assert.notNull(mainObjJoinPropertyName, "mainObjJoinPropertyName must not be null");
-
-    if (isNotEmpty(mainObjList)) {
-      Class<?> relationshipClazz =
-          getPropertyClass(mainObjList.get(0), mainObjRelationshipPropertyName);
-      TableMapping relationshipTableMapping = getTableMapping(relationshipClazz);
-      String idColumnName = relationshipTableMapping.getIdColumnName();
-      if (idColumnName == null) {
-        throw new RuntimeException(
-            "Either "
-                + relationshipClazz.getSimpleName()
-                + " does not have an 'id' property or its corresponding table "
-                + relationshipTableMapping.getTableName()
-                + " does not have an 'id' column");
-      }
-
-      LinkedHashSet<Long> allJoinPropertyIds = new LinkedHashSet<>();
-      boolean firstRecord = true;
-      for (T mainObj : mainObjList) {
-        if (mainObj != null) {
-          // Do some validations
-          if (firstRecord) {
-            firstRecord = false;
-            validateToOne(
-                mainObj,
-                relationshipClazz,
-                mainObjRelationshipPropertyName,
-                mainObjJoinPropertyName);
-          }
-
-          // create the list fo all join property ids
-          Number joinPropertyValue = (Number) getPropertyValue(mainObj, mainObjJoinPropertyName);
-          if (joinPropertyValue != null && joinPropertyValue.longValue() > 0) {
-            allJoinPropertyIds.add(joinPropertyValue.longValue());
-          }
-        }
-      }
-      List relatedObjList = new ArrayList();
-      // Since some databases have limitations on how many entries can be in an 'IN' clause and to
-      // avoid query being issued with large number of ids for the 'IN (:joinPropertyIds), the
-      // list is chunked by IN_CLAUSE_CHUNK_SIZE and multiple queries issued if needed.
-      Collection<List<Long>> chunkedJoinPropertyIds =
-          chunkTheCollection(allJoinPropertyIds, inClauseChunkSize);
-      for (List<Long> joinPropertyIds : chunkedJoinPropertyIds) {
-        String sql =
-            "SELECT * FROM "
-                + fullyQualifiedTableName(relationshipTableMapping.getTableName())
-                + " WHERE "
-                + idColumnName
-                + " IN (:joinPropertyIds)";
-        MapSqlParameterSource params =
-            new MapSqlParameterSource("joinPropertyIds", joinPropertyIds);
-        RowMapper mapper = BeanPropertyRowMapper.newInstance(relationshipClazz);
-        relatedObjList.addAll(npJdbcTemplate.query(sql, params, mapper));
-      }
-
-      toOneMerge(
-          mainObjList, relatedObjList, mainObjRelationshipPropertyName, mainObjJoinPropertyName);
-    }
-  }
-
-  /**
-   * Populates a single main object and its toOne related object with the data from the query ResultSet
-   * using their respective SqlMappers. The sql for the Resultset object should have the join
-   * properties in select statement so that the main object and related object can be tied together
-   * (mainObj.mainObjJoinPropertyName = relatedObj.id)
-   *
-   * <pre>
-   * Main Object:
-   * public class Order{
-   *   Integer id;
-   *   LocalDateTime orderDate;
-   *   Integer customerId;
-   *   Customer customer; // the toOne relationship
-   * }
-   *
-   * toOne related Object:
-   * public class Customer{
-   *   Integer id;
-   *   String firstName;
-   *   String lastName;
-   *   String address;
-   * }
-   *
-   * The sql below uses a column alias naming convention so that the SelectMapper() can use the column prefix
-   * to populate the appropriated Objects from the selected columns.
-   *
-   * Note that the join properties  o.customer_id and the c.id have to be in select clause.
-   * <pre>
-   * select o.id o_id, o.order_date o_order_date, o.customer_id o_customer_id
-   *        c.id c_id, c.first_name c_first_name, c.last_name c_last_name, c.address c_address
-   * from order o
-   * left join customer c on o.customer_id = c.id
-   * where o.id = ?
-   *
-   * See selectColumns()} on how to make the select clause less verbose.
-   *
-   * Example call to get Order and its Customer (toOne relationship) populated from the sql above:
-   * Order order =
-   *     jdbcTemplate.query(
-   *         sql,
-   *         new Object[] {theOrderId}, // args
-   *         new int[] {java.sql.Types.INTEGER}, // arg types
-   *         rs -> {
-   *           return jdbcTemplateMapper.toOneMapperForObject(
-   *               rs,
-   *               new SelectMapper<Order>(Order.class, "o_"), // maps column names with prefix 'o_' to Order
-   *               new SelectMapper<Customer>(Customer.class, "c_"), // maps column names with prefix 'c_' to Customer
-   *               "customer",
-   *               "customerId");
-   *         });
-   * </pre>
-   *
-   * @param rs The jdbc ResultSet
-   * @param mainObjMapper The main object mapper.
-   * @param relatedObjMapper The related object mapper
-   * @param mainObjRelationshipPropertyName The toOne relationship property name on main object
-   * @param mainObjJoinPropertyName The join property name on the main object
-   * @return The main object with its toOne relationship populated.
-   */
-  @SuppressWarnings("all")
-  public <T, U> T toOneMapperForObject(
-      ResultSet rs,
-      SelectMapper<T> mainObjMapper,
-      SelectMapper<U> relatedObjMapper,
-      String mainObjRelationshipPropertyName,
-      String mainObjJoinPropertyName) {
-
-    Assert.notNull(mainObjMapper, "mainObjMapper must not be null");
-    Assert.notNull(relatedObjMapper, "relatedObjMapper must not be null");
-    Assert.notNull(
-        mainObjRelationshipPropertyName, "mainObjRelationshipPropertyName must not be null");
-    Assert.notNull(mainObjJoinPropertyName, "mainObjJoinPropertyName must not be null");
-
-    List<T> list =
-        toOneMapperForList(
-            rs,
-            mainObjMapper,
-            relatedObjMapper,
-            mainObjRelationshipPropertyName,
-            mainObjJoinPropertyName);
-    return isNotEmpty(list) ? list.get(0) : null;
-  }
-
-  /**
-   * Populates the main object list with their corresponding toOne related object from the jdbc
-   * ResultSet using their respective SqlMappers. The sql for the Resultset object should have the
-   * join properties in select statement so that the main object and related object can be tied
-   * together (mainObj.mainObjJoinPropertyName = relatedObj.id)
-   *
-   * <pre>
-   * Main Object:
-   * public class Order{
-   *   Integer id;
-   *   LocalDateTime orderDate;
-   *   Integer customerId;
-   *   Customer customer; // the toOne relationship
-   * }
-   *
-   * toOne related Object:
-   * public class Customer{
-   *   Integer id;
-   *   String firstName;
-   *   String lastName;
-   *   String address;
-   * }
-   *
-   * The sql below uses a column alias naming convention so that the SelectMapper() can use the column prefix
-   * to populate the appropriated Objects from the selected columns.
-   *
-   * Note that the join properties  o.customer_id and the c.id have to be in select clause.
-   *
-   * select o.id o_id, o.order_date o_order_date, o.customer_id o_customer_id
-   *        c.id c_id, c.first_name c_first_name, c.last_name c_last_name, c.address c_address
-   * from order o
-   * left join customer c on o.customer_id = c.id
-   *
-   * See selectColumns() to make the select clause less verbose.
-   *
-   * Example call to get Orders and its Customer (toOne relationship) populated from the sql above:
-   * List<Order> orders =
-   *     jdbcTemplate.query(
-   *         sql,
-   *         rs -> {
-   *           return jdbcTemplateMapper.toOneMapperForList(
-   *               rs,
-   *               new SelectMapper<Order>(Order.class, "o_"), // maps column names with prefix 'o_' to Order
-   *               new SelectMapper<Customer>(Customer.class, "c_"), // maps column names with prefix 'c_' to Customer
-   *               "customer",
-   *               "customerId");
-   *         });
-   * </pre>
-   *
-   * @param rs The jdbc ResultSet
-   * @param mainObjMapper The main object mapper.
-   * @param relatedObjMapper The related object mapper
-   * @param mainObjRelationshipPropertyName The toOne relationship property name on main object
-   * @param mainObjJoinPropertyName The join property name on the main object
-   * @return List of mainObj with its toOne property assigned
-   */
-  @SuppressWarnings("all")
-  public <T, U> List<T> toOneMapperForList(
-      ResultSet rs,
-      SelectMapper<T> mainObjMapper,
-      SelectMapper<U> relatedObjMapper,
-      String mainObjRelationshipPropertyName,
-      String mainObjJoinPropertyName) {
-
-    Assert.notNull(mainObjMapper, "mainObjMapper must not be null");
-    Assert.notNull(relatedObjMapper, "relatedObjMapper must not be null");
-    Assert.notNull(
-        mainObjRelationshipPropertyName, "mainObjRelationshipPropertyName must not be null");
-    Assert.notNull(mainObjJoinPropertyName, "mainObjJoinPropertyName must not be null");
-
-    Map<String, LinkedHashMap<Long, Object>> resultMap =
-        multipleModelMapperRaw(rs, mainObjMapper, relatedObjMapper);
-    List<T> mainObjList = new ArrayList(resultMap.get(mainObjMapper.getSqlColumnPrefix()).values());
-    LinkedHashMap<Long, Object> relatedObjMap =
-        resultMap.get(relatedObjMapper.getSqlColumnPrefix());
-
-    // assign related obj to the main obj relationship property
-    for (Object mainObj : mainObjList) {
-      if (mainObj != null) {
-        Number joinPropertyValue = (Number) getPropertyValue(mainObj, mainObjJoinPropertyName);
-        if (joinPropertyValue != null && joinPropertyValue.longValue() > 0) {
-          setPropertyValue(
-              mainObj,
-              mainObjRelationshipPropertyName,
-              relatedObjMap.get(joinPropertyValue.longValue()));
-        }
-      }
-    }
-
-    return mainObjList;
-  }
-
-  /**
-   * Merges relatedObjectList to the mainObj list by assigning
-   * mainOjbj.mainObjRelationshipPropertyName with matching related objects ie
-   * mainObj.mainObjJoinPropertyName = relatedObj.id
-   *
-   * @param mainObjList list of main objects
-   * @param relatedObjList list of related objects
-   * @param mainObjRelationshipPropertyName The toOne relationship property name on main object
-   * @param mainObjJoinPropertyName The join property name on the main object
-   */
-  public <T, U> void toOneMerge(
-      List<T> mainObjList,
-      List<U> relatedObjList,
-      String mainObjRelationshipPropertyName,
-      String mainObjJoinPropertyName) {
-
-    Assert.notNull(
-        mainObjRelationshipPropertyName, "mainObjRelationshipPropertyName must not be null");
-    Assert.notNull(mainObjJoinPropertyName, "mainObjJoinPropertyName must not be null");
-
-    if (isNotEmpty(mainObjList) && isNotEmpty(relatedObjList)) {
-      // Map key: related object id , value: the related object
-      Map<Long, U> idToObjectMap = new HashMap<>();
-      for (U relatedObj : relatedObjList) {
-        if (relatedObj != null) {
-          Number idVal = (Number) getPropertyValue(relatedObj, "id");
-          if (idVal != null && idVal.longValue() > 0) {
-            idToObjectMap.put(idVal.longValue(), relatedObj);
-          }
-        }
-      }
-      // assign related obj to the main obj relationship property
-      for (T mainObj : mainObjList) {
-        if (mainObj != null) {
-          Number joinPropertyValue = (Number) getPropertyValue(mainObj, mainObjJoinPropertyName);
-          if (joinPropertyValue != null && joinPropertyValue.longValue() > 0) {
-            setPropertyValue(
-                mainObj,
-                mainObjRelationshipPropertyName,
-                idToObjectMap.get(joinPropertyValue.longValue()));
-          }
-        }
-      }
-    }
-  }
-
-  /**
-   * Populates the toMany relationship. Issues an sql query to get the many side records. Make sure
-   * the id property of the argument main object is assigned so that it can be matched to its
-   * corresponding many side object (mainObj.id = manySideObj.manySideJoinPropertyName)
-   *
-   * <pre>
-   * Main Object:
-   * public class Order{
-   *   Integer id;
-   *   LocalDateTime orderDate;
-   *   Integer customerId;
-   *   List<OrderLine> orderLines; // toMany relationship
-   * }
-   * toMany related Object:
-   * public class OrderLine{
-   *   Integer id;
-   *   Integer orderId;
-   *   Integer productId;
-   *   Integer quantity;
-   *   Double price;
-   * }
-   *
-   * 1) Get an Order Object using some query for example:
-   * Order order = jdbcTemplateMapper.findById(orderId)
-   *
-   * 2) Populate the order's orderLines. This will issue an sql and populate order.orderLines
-   * jdbcTemplateMapper.toMany(order, "orderLines", "orderId");
-   *
-   * </pre>
-   *
-   * @param mainObjList the main object list
-   * @param mainObjCollectionPropertyName The collection property name on main object
-   * @param manySideJoinPropertyName The join property name on the many side object
-   */
-  public <T> void toMany(
-      T mainObj, String mainObjCollectionPropertyName, String manySideJoinPropertyName) {
-
-    Assert.notNull(mainObjCollectionPropertyName, "mainObjCollectionPropertyName must not be null");
-    Assert.notNull(manySideJoinPropertyName, "manySideJoinPropertyName must not be null");
-
-    toMany(mainObj, mainObjCollectionPropertyName, manySideJoinPropertyName, null);
-  }
-
-  /**
-   * Populates the toMany relationship. Issues an sql query to get the many side records. The many
-   * side ordering can be customized using the last argument. Make sure the id property of the
-   * argument main object is assigned so that it can be matched to its corresponding many side
-   * object (mainObj.id = manySideObj.manySideJoinPropertyName)
-   *
-   * <pre>
-   * Main Object:
-   * public class Order{
-   *   Integer id;
-   *   LocalDateTime orderDate;
-   *   Integer customerId;
-   *   List<OrderLine> orderLines; // toMany relationship
-   * }
-   * toMany related Object:
-   * public class OrderLine{
-   *   Integer id;
-   *   Integer orderId;
-   *   Integer productId;
-   *   Integer quantity;
-   *   Double price;
-   * }
-   *
-   * 1) Get an Order Object using some query for example:
-   * Order order = jdbcTemplateMapper.findById(orderId)
-   *
-   * 2) Populate the order's orderLines. This will issue an sql with the ordering clause and populate order.orderLines
-   * jdbcTemplateMapper.toMany(order, "orderLines", "orderId", "order by price");
-   * </pre>
-   *
-   * @param mainObjList the main object list
-   * @param mainObjCollectionPropertyName The collection property name on main object
-   * @param manySideJoinPropertyName The join property name on the many side object
-   * @param manySideOrderByClause The order by clause for the many side query
-   */
-  public <T, U> void toMany(
-      T mainObj,
-      String mainObjCollectionPropertyName,
-      String manySideJoinPropertyName,
-      String manySideOrderByClause) {
-
-    Assert.notNull(mainObjCollectionPropertyName, "mainObjCollectionPropertyName must not be null");
-    Assert.notNull(manySideJoinPropertyName, "manySideJoinPropertyName must not be null");
-
-    if (mainObj != null) {
-      List<T> mainObjList = new ArrayList<>();
-      mainObjList.add(mainObj);
-      toMany(
-          mainObjList,
-          mainObjCollectionPropertyName,
-          manySideJoinPropertyName,
-          manySideOrderByClause);
-    }
-  }
-
-  /**
-   * Populates the toMany relationship of each the main objects in the list. Issues an sql query to
-   * get the many side records. Make sure the id property of the main objects are assigned so that
-   * they can be matched to their corresponding many side objects (mainObj.id =
-   * manySideObj.manySideJoinPropertyName)
-   *
-   * <pre>
-   * Main Object:
-   * public class Order{
-   *   Integer id;
-   *   LocalDateTime orderDate;
-   *   Integer customerId;
-   *   List<OrderLine> orderLines; // toMany relationship
-   * }
-   * toMany related Object:
-   * public class OrderLine{
-   *   Integer id;
-   *   Integer orderId;
-   *   Integer productId;
-   *   Integer quantity;
-   *   Double price;
-   * }
-   *
-   * 1) Get a list of Order objects using some query for example:
-   * List<Order> orders = jdbcTemplateMapper.findAll(Order.class);
-   *
-   * 2) Populate each order's orderlines . This will issue an sql (with an IN clause) and
-   *    populate order.orderLines.
-   * jdbcTemplateMapper.toMany(orders, "orderLines", "orderId");
-   *
-   * </pre>
-   *
-   * @param mainObjList the main object list
-   * @param mainObjCollectionPropertyName The collection property name on mainObj
-   * @param manySideJoinPropertyName the join property name on the many side object
-   */
-  public <T> void toMany(
-      List<T> mainObjList, String mainObjCollectionPropertyName, String manySideJoinPropertyName) {
-
-    Assert.notNull(mainObjCollectionPropertyName, "mainObjCollectionPropertyName must not be null");
-    Assert.notNull(manySideJoinPropertyName, "manySideJoinPropertyName must not be null");
-
-    toMany(mainObjList, mainObjCollectionPropertyName, manySideJoinPropertyName, null);
-  }
-
-  /**
-   * Populates the toMany relations of the list of main objects. Issues an sql query to get the many
-   * side records. The many side ordering can be customized using the last argument. Make sure the
-   * id property of the main objects are assigned so that they can be matched to their corresponding
-   * many side objects (mainObj.id = manySideObj.manySideJoinPropertyName)
-   *
-   * <pre>
-   * Main Object:
-   * public class Order{
-   *   Integer id;
-   *   LocalDateTime orderDate;
-   *   Integer customerId;
-   *   List<OrderLine> orderLines; // toMany relationship
-   * }
-   * toMany related Object:
-   * public class OrderLine{
-   *   Integer id;
-   *   Integer orderId;
-   *   Integer productId;
-   *   Integer quantity;
-   *   Double price;
-   * }
-   *
-   * 1) Get a list of Order objects using some query for example:
-   * List<Order> orders = jdbcTemplateMapper.findAll(Order.class);
-   *
-   * 2) Populate each Order's orderLines . This will issue an sql with the ordering clause and populate order.orderLines.
-   * jdbcTemplateMapper.toMany(orders, "orderLines", "orderId", "order by price");
-   *
-   * </pre>
-   *
-   * @param mainObjList the main object list
-   * @param mainObjCollectionPropertyName The collection property name on mainObj
-   * @param manySideJoinPropertyName the join property name on the many side object
-   * @param manySideOrderByClause The order by clause for the many side query. Use the database
-   *     columns in this clause because it is passed as is.
-   */
-  @SuppressWarnings("all")
-  public <T> void toMany(
-      List<T> mainObjList,
-      String mainObjCollectionPropertyName,
-      String manySideJoinPropertyName,
-      String manySideOrderByClause) {
-
-    Assert.notNull(mainObjCollectionPropertyName, "mainObjCollectionPropertyName must not be null");
-    Assert.notNull(manySideJoinPropertyName, "manySideJoinPropertyName must not be null");
-
-    if (isNotEmpty(mainObjList)) {
-      Class<?> manySideClazz =
-          getGenericTypeOfCollection(mainObjList.get(0), mainObjCollectionPropertyName);
-      String manySideTableName = getTableMapping(manySideClazz).getTableName();
-      LinkedHashSet<Long> allMainObjIds = new LinkedHashSet<>();
-      boolean firstRecord = true;
-      for (T mainObj : mainObjList) {
-        if (mainObj != null) {
-          if (firstRecord) {
-            validateToManyMainObj(mainObj, mainObjCollectionPropertyName);
-            validateToManyManySideObj(manySideClazz, manySideJoinPropertyName);
-            firstRecord = false;
-          }
-
-          Number idVal = (Number) getPropertyValue(mainObj, "id");
-          if (idVal != null && idVal.longValue() > 0) {
-            allMainObjIds.add(idVal.longValue());
-          } else {
-            throw new RuntimeException("id property in mainObjList cannot be null");
-          }
-        }
-      }
-
-      String joinColumnName = getJoinColumnName(manySideTableName, manySideJoinPropertyName);
-      List manySideList = new ArrayList();
-      // Since some databases have limitations on how many entries can be in an 'IN' clause and to
-      // avoid
-      // query being issued with large number of ids for the 'IN (:mainObjIds), the list
-      // is chunked by IN_CLAUSE_CHUNK_SIZE and multiple queries issued if needed.
-      Collection<List<Long>> chunkedMainObjIds =
-          chunkTheCollection(allMainObjIds, inClauseChunkSize);
-      for (List<Long> mainObjIds : chunkedMainObjIds) {
-        String sql =
-            "SELECT * FROM "
-                + fullyQualifiedTableName(manySideTableName)
-                + " WHERE "
-                + joinColumnName
-                + " IN (:mainObjIds)";
-        if (isNotEmpty(manySideOrderByClause)) {
-          sql += " " + manySideOrderByClause;
-        }
-        MapSqlParameterSource params = new MapSqlParameterSource("mainObjIds", mainObjIds);
-        RowMapper mapper = BeanPropertyRowMapper.newInstance(manySideClazz);
-        manySideList.addAll(npJdbcTemplate.query(sql, params, mapper));
-      }
-
-      toManyMerge(
-          mainObjList, manySideList, mainObjCollectionPropertyName, manySideJoinPropertyName);
-    }
-  }
-
-  /**
-   * Populates a single main object and its many side collection with the data from the ResultSet
-   * using their respective SqlMappers.
-   *
-   * <p>The jdbc ResultSet should have mainObj.id and manySideObj.manySideJoinPropertyName so that
-   * the mainObj.mainObjCollectionProperty can be assigned by matching mainObj.id to
-   * manySideObj.manySideJoinPropertyName.
-   *
-   * @param rs The jdbc ResultSet
-   * @param mainObjMapper The main object mapper.
-   * @param manySideObjMapper The many side object mapper
-   * @param mainObjCollectionPropertyName the collectionPropertyName on the mainObj that needs to be
-   *     populated
-   * @param manySideJoinPropertyName the join property name on the manySide
-   * @return The main object with its toMany relationship assigned.
-   */
-  @SuppressWarnings("all")
-  public <T, U> T toManyMapperForObject(
-      ResultSet rs,
-      SelectMapper<T> mainObjMapper,
-      SelectMapper<U> manySideObjMapper,
-      String mainObjCollectionPropertyName,
-      String manySideJoinPropertyName) {
-
-    Assert.notNull(mainObjMapper, "mainObjMapper must not be null");
-    Assert.notNull(manySideObjMapper, "manySideObjMapper must not be null");
-    Assert.notNull(mainObjCollectionPropertyName, "mainObjCollectionPropertyName must not be null");
-    Assert.notNull(manySideJoinPropertyName, "manySideJoinPropertyName must not be null");
-
-    List<T> list =
-        toManyMapperForList(
-            rs,
-            mainObjMapper,
-            manySideObjMapper,
-            mainObjCollectionPropertyName,
-            manySideJoinPropertyName);
-    return isNotEmpty(list) ? list.get(0) : null;
-  }
-
-  /**
-   * Populates the main object list and their corresponding many side collections from the jdbc
-   * ResultSet using their respective SqlMappers.
-   *
-   * <p>The jdbc ResultSet should have mainObj.id and manySideObj.manySideJoinPropertyName so that
-   * the mainObj.mainObjCollectionProperty can be assigned by matching mainObj.id to
-   * manySideObj.manySideJoinPropertyName.
-   *
-   * @param rs The jdbc ResultSet
-   * @param mainObjMapper The main object mapper
-   * @param manySideObjMapper The many side object mapper
-   * @param mainObjCollectionPropertyName the collectionPropertyName on the mainObj that needs to be
-   *     populated
-   * @param manySideJoinPropertyName the join property name on the manySide
-   * @return List of mainObj with its collectionPropertyName populated
-   */
-  @SuppressWarnings("all")
-  public <T, U> List<T> toManyMapperForList(
-      ResultSet rs,
-      SelectMapper<T> mainObjMapper,
-      SelectMapper<U> manySideObjMapper,
-      String mainObjCollectionPropertyName,
-      String manySideJoinPropertyName) {
-
-    Assert.notNull(mainObjMapper, "mainObjMapper must not be null");
-    Assert.notNull(manySideObjMapper, "manySideObjMapper must not be null");
-    Assert.notNull(mainObjCollectionPropertyName, "mainObjCollectionPropertyName must not be null");
-    Assert.notNull(manySideJoinPropertyName, "manySideJoinPropertyName must not be null");
-
-    Map<String, List> resultMap = multipleModelMapper(rs, mainObjMapper, manySideObjMapper);
-    List<T> mainObjList = resultMap.get(mainObjMapper.getSqlColumnPrefix());
-    List<U> manySideObjList = resultMap.get(manySideObjMapper.getSqlColumnPrefix());
-
-    toManyMerge(
-        mainObjList, manySideObjList, mainObjCollectionPropertyName, manySideJoinPropertyName);
-    return mainObjList;
-  }
-
-  /**
-   * Populates each main objects collectionPropertyName with the corresponding manySide objects by
-   * matching manySide.joinPropertyName to mainObj.id
-   *
-   * @param mainObjList the main object list
-   * @param manySideList the many side object list
-   * @param mainObjCollectionPropertyName the collection property name of the main object that needs
-   *     to be populated.
-   * @param manySideJoinPropertyName the join property name on the many side
-   */
-  @SuppressWarnings("all")
-  public <T, U> void toManyMerge(
-      List<T> mainObjList,
-      List<U> manySideList,
-      String mainObjCollectionPropertyName,
-      String manySideJoinPropertyName) {
-
-    Assert.notNull(mainObjCollectionPropertyName, "mainObjCollectionPropertyName must not be null");
-    Assert.notNull(manySideJoinPropertyName, "manySideJoinPropertyName must not be null");
-
-    try {
-      if (isNotEmpty(mainObjList) && isNotEmpty(manySideList)) {
-        // many side records are grouped by their join property values
-        // Map key - join property value , value - List of records grouped by the join property
-        // value
-        Map<Long, List<U>> groupedManySide = new HashMap<>();
-        for (U manySideObj : manySideList) {
-          if (manySideObj != null) {
-            Number joinPropertyValue =
-                (Number) getPropertyValue(manySideObj, manySideJoinPropertyName);
-            if (joinPropertyValue != null && joinPropertyValue.longValue() > 0) {
-              if (groupedManySide.containsKey(joinPropertyValue.longValue())) {
-                groupedManySide.get(joinPropertyValue.longValue()).add(manySideObj);
-              } else {
-                List<U> list = new ArrayList<>();
-                list.add(manySideObj);
-                groupedManySide.put(joinPropertyValue.longValue(), list);
-              }
-            }
-          }
-        }
-        // assign the manyside list to the mainobj
-        for (T mainObj : mainObjList) {
-          if (mainObj != null) {
-            Number idVal = (Number) getPropertyValue(mainObj, "id");
-            if (idVal != null && idVal.longValue() > 0) {
-              setPropertyValue(
-                  mainObj, mainObjCollectionPropertyName, groupedManySide.get(idVal.longValue()));
-            }
-          }
-        }
-      }
-    } catch (Exception e) {
-      throw new RuntimeException(e);
-    }
-  }
 
   /**
    * Returns lists for each mapper passed in as an argument. The values in the list are UNIQUE and
@@ -1558,7 +761,7 @@ public class JdbcTemplateMapper {
       for (SelectMapper selectMapper : selectMappers) {
         resultMap.put(selectMapper.getSqlColumnPrefix(), new LinkedHashMap<>());
       }
-      List<String> resultSetColumnNames = getResultSetColumnNames(rs);
+      List<String> resultSetColumnNames = dbUtils.getResultSetColumnNames(rs);
       while (rs.next()) {
         for (SelectMapper selectMapper : selectMappers) {
           Number idVal = (Number) rs.getObject(selectMapper.getSqlColumnPrefix() + "id");
@@ -1599,11 +802,11 @@ public class JdbcTemplateMapper {
 
     String str = selectColumnsCache.get(tableName + "-" + tableAlias);
     if (str == null) {
-      List<ColumnInfo> tableColumnInfo = getTableColumnInfo(tableName);
-      if (isEmpty(tableColumnInfo)) {
+      List<ColumnInfo> tableColumnInfo = dbUtils.getTableColumnInfo(tableName);
+      if (CommonUtils.isEmpty(tableColumnInfo)) {
         // try with uppercase table name
-        tableColumnInfo = getTableColumnInfo(tableName.toUpperCase());
-        if (isEmpty(tableColumnInfo)) {
+        tableColumnInfo = dbUtils.getTableColumnInfo(tableName.toUpperCase());
+        if (CommonUtils.isEmpty(tableColumnInfo)) {
           throw new RuntimeException("Could not get column info for table named " + tableName);
         }
       }
@@ -1638,7 +841,7 @@ public class JdbcTemplateMapper {
     }
     Set<String> params = new HashSet<>();
     StringBuilder sqlBuilder = new StringBuilder("UPDATE ");
-    sqlBuilder.append(fullyQualifiedTableName(tableMapping.getTableName()));
+    sqlBuilder.append(dbUtils.fullyQualifiedTableName(tableMapping.getTableName()));
     sqlBuilder.append(" SET ");
 
     String versionColumnName = tableMapping.getColumnName(versionPropertyName);
@@ -1763,9 +966,9 @@ public class JdbcTemplateMapper {
       BeanWrapper bw = PropertyAccessorFactory.forBeanPropertyAccess(obj);
       // need below for java.sql.Timestamp to java.time.LocalDateTime conversion etc
       bw.setConversionService(defaultConversionService);
-      for (PropertyInfo propertyInfo : getObjectPropertyInfo(obj)) {
-        String columnName = convertCamelToSnakeCase(propertyInfo.getPropertyName());
-        if (isNotEmpty(prefix)) {
+      for (PropertyInfo propertyInfo : CommonUtils.getObjectPropertyInfo(obj)) {
+        String columnName = CommonUtils.convertCamelToSnakeCase(propertyInfo.getPropertyName());
+        if (CommonUtils.isNotEmpty(prefix)) {
           columnName = prefix + columnName;
         }
         int index = resultSetColumnNames.indexOf(columnName.toLowerCase());
@@ -1783,523 +986,13 @@ public class JdbcTemplateMapper {
   }
 
   /**
-   * Validates the arguments for toOne() methods
-   *
-   * @param mainObj The main Object
-   * @param relationshipClazz The relationship class
-   * @param mainObjRelationshipPropertyName The main object relationship property name
-   * @param mainObjJoinPropertyName The main object join property name
-   */
-  private void validateToOne(
-      Object mainObj,
-      Class<?> relationshipClazz,
-      String mainObjRelationshipPropertyName,
-      String mainObjJoinPropertyName) {
-    List<PropertyInfo> propertyInfoList = getObjectPropertyInfo(mainObj);
-
-    if ("id".equals(mainObjRelationshipPropertyName)) {
-      throw new RuntimeException("The argument mainObjRelationshipPropertyName cannot be 'id'.");
-    }
-
-    // validate mainObjRelationshipPropertyName
-    PropertyInfo propertyInfo =
-        propertyInfoList
-            .stream()
-            .filter(pi -> mainObjRelationshipPropertyName.equals(pi.getPropertyName()))
-            .findAny()
-            .orElse(null);
-
-    if (propertyInfo == null) {
-      throw new RuntimeException(
-          "property "
-              + mainObjRelationshipPropertyName
-              + " does not exist for object "
-              + mainObj.getClass().getName());
-    } else {
-      if (!relationshipClazz.equals(propertyInfo.getPropertyType())) {
-        throw new RuntimeException(
-            "type mismatch. property "
-                + mainObjRelationshipPropertyName
-                + " is of type "
-                + propertyInfo.getPropertyType()
-                + " while the argment relationshipClazz is "
-                + relationshipClazz.getName());
-      }
-    }
-
-    // validate mainObjJoinPropertyName
-    if ("id".equals(mainObjJoinPropertyName)) {
-      throw new RuntimeException("The argument mainObjJoinPropertyName cannot be 'id'.");
-    }
-
-    propertyInfo =
-        propertyInfoList
-            .stream()
-            .filter(pi -> mainObjJoinPropertyName.equals(pi.getPropertyName()))
-            .findAny()
-            .orElse(null);
-
-    if (propertyInfo == null) {
-      throw new RuntimeException(
-          "property " + mainObjJoinPropertyName + " not found in object " + mainObj);
-    } else {
-      if (!Integer.class.equals(propertyInfo.getPropertyType())
-          && !Long.class.equals(propertyInfo.getPropertyType())) {
-        throw new RuntimeException(
-            "type of property "
-                + mainObjJoinPropertyName
-                + " which is used as a join property has to be of type Integer or Long ");
-      }
-    }
-  }
-
-  /**
-   * Validates the main object info for the toMany() methods
-   *
-   * @param mainObj
-   * @param mainObjCollectionPropertyName
-   */
-  private void validateToManyMainObj(Object mainObj, String mainObjCollectionPropertyName) {
-
-    List<PropertyInfo> propertyInfoList = getObjectPropertyInfo(mainObj);
-
-    if ("id".equals(mainObjCollectionPropertyName)) {
-      throw new RuntimeException("The argument mainObjRelationshipPropertyName cannot be 'id'.");
-    }
-
-    // validate mainObjRelationshipPropertyName
-    PropertyInfo propertyInfo =
-        propertyInfoList
-            .stream()
-            .filter(pi -> mainObjCollectionPropertyName.equals(pi.getPropertyName()))
-            .findAny()
-            .orElse(null);
-
-    if (propertyInfo == null) {
-      throw new RuntimeException(
-          "property "
-              + mainObjCollectionPropertyName
-              + " does not exist for object "
-              + mainObj.getClass().getName());
-    } else {
-      if (!List.class.equals(propertyInfo.getPropertyType())) {
-        throw new RuntimeException(
-            "property " + mainObjCollectionPropertyName + " should be of type List.");
-      }
-    }
-  }
-
-  /**
-   * Validates the many side object args for toMany() methods
-   *
-   * @param manySideClazz
-   * @param manySideJoinPropertyName
-   */
-  private void validateToManyManySideObj(Class<?> manySideClazz, String manySideJoinPropertyName) {
-    List<PropertyInfo> propertyInfoList = null;
-    try {
-      propertyInfoList = getObjectPropertyInfo(manySideClazz.newInstance());
-    } catch (Exception e) {
-      throw new RuntimeException(e);
-    }
-
-    if ("id".equals(manySideJoinPropertyName)) {
-      throw new RuntimeException("The argument manySideJoinPropertyName cannot be 'id'.");
-    }
-
-    PropertyInfo propertyInfo =
-        propertyInfoList
-            .stream()
-            .filter(pi -> manySideJoinPropertyName.equals(pi.getPropertyName()))
-            .findAny()
-            .orElse(null);
-
-    if (propertyInfo == null) {
-      throw new RuntimeException(
-          "property "
-              + manySideJoinPropertyName
-              + " does not exist for object "
-              + manySideClazz.getName());
-    } else {
-      if (!Integer.class.equals(propertyInfo.getPropertyType())
-          && !Long.class.equals(propertyInfo.getPropertyType())) {
-        throw new RuntimeException(
-            "type of property "
-                + manySideJoinPropertyName
-                + " which is used as a join property has to be of type Integer or Long ");
-      }
-    }
-  }
-
-  /**
-   * Converts an object to a map with key as database column names and values the corresponding
-   * object value. Camel case property names are converted to snake case. For example property name
-   * 'userLastName' will get converted to map key 'user_last_name' and assigned the corresponding
-   * object value.
-   *
-   * @param obj The object to convert
-   * @return A map with keys that are in snake case to match database column names and values
-   *     corresponding to the object property
-   */
-  private Map<String, Object> convertToSnakeCaseAttributes(Object obj) {
-    Assert.notNull(obj, "Object must not be null");
-
-    Map<String, Object> camelCaseAttrs = convertObjectToMap(obj);
-    Map<String, Object> snakeCaseAttrs = new HashMap<>();
-    for (String key : camelCaseAttrs.keySet()) {
-      // ex: lastName will get converted to last_name
-      String snakeCaseKey = convertCamelToSnakeCase(key);
-      snakeCaseAttrs.put(snakeCaseKey, camelCaseAttrs.get(key));
-    }
-    return snakeCaseAttrs;
-  }
-
-  /**
-   * Converts an object to a Map. The map key will be object property name and value with
-   * corresponding object property value.
-   *
-   * @param obj The object to be converted.
-   * @return Map with key: property name, value: object value
-   */
-  private Map<String, Object> convertObjectToMap(Object obj) {
-    Assert.notNull(obj, "Object must not be null");
-    Map<String, Object> camelCaseAttrs = new HashMap<>();
-    BeanWrapper bw = PropertyAccessorFactory.forBeanPropertyAccess(obj);
-    for (PropertyInfo propertyInfo : getObjectPropertyInfo(obj)) {
-      camelCaseAttrs.put(
-          propertyInfo.getPropertyName(), bw.getPropertyValue(propertyInfo.getPropertyName()));
-    }
-    return camelCaseAttrs;
-  }
-
-  private List<ColumnInfo> getTableColumnInfo(String tableName) {
-    Assert.hasLength(tableName, "tableName must not be empty");
-    try {
-      List<ColumnInfo> columnInfos = tableColumnInfoCache.get(tableName);
-      if (isEmpty(columnInfos)) {
-        // Using Spring JdbcUtils.extractDatabaseMetaData() since it has some robust processing for
-        // connection access
-        columnInfos =
-            JdbcUtils.extractDatabaseMetaData(
-                jdbcTemplate.getDataSource(),
-                new DatabaseMetaDataCallback<List<ColumnInfo>>() {
-                  public List<ColumnInfo> processMetaData(DatabaseMetaData metadata)
-                      throws SQLException, MetaDataAccessException {
-                    ResultSet rs = null;
-                    try {
-                      List<ColumnInfo> columnInfoList = new ArrayList<>();
-                      rs =
-                          metadata.getColumns(
-                              catalogName, schemaName, tableName, metaDataColumnNamePattern);
-                      while (rs.next()) {
-                        columnInfoList.add(
-                            new ColumnInfo(rs.getString("COLUMN_NAME"), rs.getInt("DATA_TYPE")));
-                      }
-                      if (isNotEmpty(columnInfoList)) {
-                        tableColumnInfoCache.put(tableName, columnInfoList);
-                      }
-                      return columnInfoList;
-                    } finally {
-                      JdbcUtils.closeResultSet(rs);
-                    }
-                  }
-                });
-      }
-      return columnInfos;
-    } catch (Exception e) {
-      throw new RuntimeException();
-    }
-  }
-
-  /**
-   * Gets the resultSet lower case column names ie the column names in the 'select' statement of the
-   * sql
-   *
-   * @param rs The jdbc ResultSet
-   * @return List of select column names in lower case
-   */
-  private List<String> getResultSetColumnNames(ResultSet rs) {
-    List<String> rsColNames = new ArrayList<>();
-    try {
-      ResultSetMetaData rsmd = rs.getMetaData();
-      int numberOfColumns = rsmd.getColumnCount();
-      // jdbc indexes start at 1
-      for (int i = 1; i <= numberOfColumns; i++) {
-        rsColNames.add(rsmd.getColumnLabel(i).toLowerCase());
-      }
-    } catch (Exception e) {
-      throw new RuntimeException(e);
-    }
-    return rsColNames;
-  }
-
-  /**
-   * Get property information of an object. The property infos are cached by the object class name
-   *
-   * @param obj The object
-   * @return List of PropertyInfo
-   */
-  private List<PropertyInfo> getObjectPropertyInfo(Object obj) {
-    Assert.notNull(obj, "Object must not be null");
-    List<PropertyInfo> propertyInfoList = objectPropertyInfoCache.get(obj.getClass().getName());
-    if (propertyInfoList == null) {
-      BeanWrapper bw = PropertyAccessorFactory.forBeanPropertyAccess(obj);
-      propertyInfoList = new ArrayList<>();
-      PropertyDescriptor[] propertyDescriptors = bw.getPropertyDescriptors();
-      for (PropertyDescriptor pd : propertyDescriptors) {
-        String propName = pd.getName();
-        // log.debug("Property name:{}" + propName);
-        if ("class".equals(propName)) {
-          continue;
-        } else {
-          propertyInfoList.add(new PropertyInfo(propName, pd.getPropertyType()));
-        }
-      }
-      objectPropertyInfoCache.put(obj.getClass().getName(), propertyInfoList);
-    }
-    return propertyInfoList;
-  }
-
-  /**
-   * Get a property using reflection. Used instead of BeanWrapper since it is too heavy creating an
-   * instance of BeanWrapper to just lookup 1 property. The getter Methods are cached for
-   * performance.
-   *
-   * @param obj The object
-   * @param propertyName The property that needs needs to be retrieved
-   * @return the value of property
-   */
-  private Object getPropertyValue(Object obj, String propertyName) {
-    Assert.notNull(obj, "Object must not be null");
-    Assert.hasLength(propertyName, "propertyName must not be empty");
-
-    Method method = objectGetMethodCache.get(obj.getClass().getName() + "-" + propertyName);
-    if (method == null) {
-      method =
-          ReflectionUtils.findMethod(obj.getClass(), "get" + StringUtils.capitalize(propertyName));
-      if (method == null) {
-        throw new RuntimeException(
-            "method get"
-                + StringUtils.capitalize(propertyName)
-                + "() not found in object "
-                + obj.getClass().getName());
-      }
-      objectGetMethodCache.put(obj.getClass().getName() + "-" + propertyName, method);
-    }
-    return ReflectionUtils.invokeMethod(method, obj);
-  }
-
-  /**
-   * Set a property using reflection. Used instead of BeanWrapper since it is too heavy creating an
-   * instance of BeanWrapper just to set 1 property. The setter Methods are cached for performance.
-   *
-   * @param obj The object
-   * @param propertyName The property that needs to be set
-   * @param val The value that needs to be set
-   */
-  private void setPropertyValue(Object obj, String propertyName, Object val) {
-    Assert.notNull(obj, "Object must not be null");
-    Assert.hasLength(propertyName, "propertyName must not be empty");
-
-    Method method = objectSetMethodCache.get(obj.getClass().getName() + "-" + propertyName);
-    if (method == null) {
-      Field field = ReflectionUtils.findField(obj.getClass(), propertyName);
-      if (field == null) {
-        throw new RuntimeException(
-            "property " + propertyName + " not found in class " + obj.getClass().getName());
-      }
-      method =
-          ReflectionUtils.findMethod(
-              obj.getClass(), "set" + StringUtils.capitalize(propertyName), field.getType());
-      if (method == null) {
-        throw new RuntimeException(
-            "method set"
-                + StringUtils.capitalize(propertyName)
-                + "("
-                + field.getType().getName()
-                + ") not found in object "
-                + obj.getClass().getName());
-      }
-      objectSetMethodCache.put(obj.getClass().getName() + "-" + propertyName, method);
-    }
-    ReflectionUtils.invokeMethod(method, obj, val);
-  }
-
-  /**
-   * Gets the table mapping for the Object. The table mapping has the table name and and object
-   * property to database column mapping.
-   *
-   * <p>Table name is either from the @Tabel annotation or the snake case conversion of the Object
-   * name.
-   *
-   * @param clazz The object class
-   * @return The table mapping.
-   */
-  private TableMapping getTableMapping(Class<?> clazz) {
-    Assert.notNull(clazz, "Class must not be null");
-    TableMapping tableMapping = objectToTableMappingCache.get(clazz.getName());
-
-    if (tableMapping == null) {
-      String tableName = null;
-      Table tableAnnotation = AnnotationUtils.findAnnotation(clazz, Table.class);
-      if (tableAnnotation != null) {
-        tableName = tableAnnotation.name();
-      } else {
-        tableName = convertCamelToSnakeCase(clazz.getSimpleName());
-      }
-
-      List<ColumnInfo> columnInfoList = getTableColumnInfo(tableName);
-      if (isEmpty(columnInfoList)) {
-        // try again with upper case table name
-        tableName = tableName.toUpperCase();
-        columnInfoList = getTableColumnInfo(tableName);
-        if (isEmpty(columnInfoList)) {
-          throw new RuntimeException(
-              "Could not find corresponding table for class " + clazz.getSimpleName());
-        }
-      }
-
-      // if code reaches here table exists
-      try {
-        List<PropertyInfo> propertyInfoList = getObjectPropertyInfo(clazz.newInstance());
-        List<PropertyMapping> propertyMappings = new ArrayList<>();
-        // Match  database table columns to the Object properties
-        for (ColumnInfo columnInfo : columnInfoList) {
-          // property name corresponding to column name
-          String propertyName = convertSnakeToCamelCase(columnInfo.getColumnName());
-          // check if the property exists for the Object
-          PropertyInfo propertyInfo =
-              propertyInfoList
-                  .stream()
-                  .filter(pi -> propertyName.equals(pi.getPropertyName()))
-                  .findAny()
-                  .orElse(null);
-
-          // add matched object property info and table column info to mappings
-          if (propertyInfo != null) {
-            propertyMappings.add(
-                new PropertyMapping(
-                    propertyInfo.getPropertyName(),
-                    propertyInfo.getPropertyType(),
-                    columnInfo.getColumnName(),
-                    columnInfo.getColumnSqlDataType()));
-          }
-        }
-
-        tableMapping = new TableMapping();
-        tableMapping.setTableName(tableName);
-        tableMapping.setPropertyMappings(propertyMappings);
-
-        // get the case sensitive id name
-        for (ColumnInfo columnInfo : columnInfoList) {
-          if ("id".equals(columnInfo.getColumnName()) || "ID".equals(columnInfo.getColumnName())) {
-            tableMapping.setIdColumnName(columnInfo.getColumnName());
-            break;
-          }
-        }
-      } catch (Exception e) {
-        throw new RuntimeException(e);
-      }
-    }
-    objectToTableMappingCache.put(clazz.getName(), tableMapping);
-    return tableMapping;
-  }
-
-  private String getJoinColumnName(String tableName, String joinPropertyName) {
-    Assert.hasLength(tableName, "tableName must not be empty");
-    Assert.hasLength(joinPropertyName, "joinPropertyName must not be empty");
-
-    List<ColumnInfo> columnInfoList = getTableColumnInfo(tableName);
-
-    String joinColumnName = convertCamelToSnakeCase(joinPropertyName);
-    String ucJoinColumnName = joinColumnName.toUpperCase();
-    for (ColumnInfo columnInfo : columnInfoList) {
-      if (joinColumnName.equals(columnInfo.getColumnName())
-          || ucJoinColumnName.equals(columnInfo.getColumnName())) {
-        return columnInfo.getColumnName();
-      }
-    }
-    // if code reached here throw exception
-    throw new RuntimeException(
-        "Could not find corresponding column in table "
-            + tableName
-            + " for joinPropertyName "
-            + joinPropertyName);
-  }
-
-  /**
-   * Converts camel case to snake case. Ex: userLastName gets converted to user_last_name. The
-   * conversion info is cached.
-   *
-   * @param str camel case String
-   * @return the snake case string to lower case.
-   */
-  private String convertCamelToSnakeCase(String str) {
-    String snakeCase = camelToSnakeCache.get(str);
-    if (snakeCase == null) {
-      if (str != null) {
-        snakeCase = TO_SNAKE_CASE_PATTERN.matcher(str).replaceAll("$1_$2").toLowerCase();
-        camelToSnakeCache.put(str, snakeCase);
-      }
-    }
-    return snakeCase;
-  }
-
-  /**
-   * Converts snake case to camel case. Ex: user_last_name gets converted to userLastName. The
-   * conversion info is cached.
-   *
-   * @param str snake case string
-   * @return the camel case string
-   */
-  private String convertSnakeToCamelCase(String str) {
-    return JdbcUtils.convertUnderscoreNameToPropertyName(str);
-  }
-
-  /**
-   * Splits the list into multiple lists of chunk size. Used to split the sql IN clauses since some
-   * databases have a limitation of 1024.
-   *
-   * @param list The list of Long
-   * @param chunkSize The size of each chunk
-   * @return Collection of lists broken down by chunkSize
-   */
-  public Collection<List<Long>> chunkTheCollection(
-      Collection<Long> collection, Integer chunkSize) {
-    Assert.notNull(collection, "collection must not be null");
-    AtomicInteger counter = new AtomicInteger();
-    Collection<List<Long>> result =
-        collection
-            .stream()
-            .filter(e -> e != null)
-            .collect(Collectors.groupingBy(it -> counter.getAndIncrement() / chunkSize))
-            .values();
-    return result;
-  }
-
-  /**
-   * Get the fully qualified table name. If schema is present then will be schemaName.tableName
-   * otherwise just tableName
-   *
-   * @param tableName The table name
-   * @return The fully qualified table name
-   */
-  private String fullyQualifiedTableName(String tableName) {
-    Assert.hasLength(tableName, "tableName must not be empty");
-    if (isNotEmpty(schemaName)) {
-      return schemaName + "." + tableName;
-    }
-    return tableName;
-  }
-
-  /**
    * checks if an object has an property named 'id'
    *
    * @param obj The object
    * @return true/false depending on if obj has 'id' property
    */
   private boolean hasIdProperty(Object obj) {
-    List<PropertyInfo> propertyInfoList = getObjectPropertyInfo(obj);
+    List<PropertyInfo> propertyInfoList = CommonUtils.getObjectPropertyInfo(obj);
     PropertyInfo propertyInfo =
         propertyInfoList
             .stream()
@@ -2308,67 +1001,5 @@ public class JdbcTemplateMapper {
             .orElse(null);
 
     return propertyInfo == null ? false : true;
-  }
-
-  private boolean isEmpty(String str) {
-    return str == null || str.length() == 0;
-  }
-
-  private boolean isNotEmpty(String str) {
-    return !isEmpty(str);
-  }
-
-  @SuppressWarnings("all")
-  private boolean isEmpty(Collection coll) {
-    return (coll == null || coll.isEmpty());
-  }
-
-  @SuppressWarnings("all")
-  private boolean isNotEmpty(Collection coll) {
-    return !isEmpty(coll);
-  }
-
-  private Class<?> getGenericTypeOfCollection(Object mainObj, String propertyName) {
-    Assert.notNull(mainObj, "mainObj must not be null");
-    Assert.notNull(propertyName, "propertyName must not be null");
-    try {
-      Field field = mainObj.getClass().getDeclaredField(propertyName);
-
-      ParameterizedType pt = (ParameterizedType) field.getGenericType();
-      Type[] genericType = pt.getActualTypeArguments();
-
-      if (genericType != null && genericType.length > 0) {
-        return Class.forName(genericType[0].getTypeName());
-      } else {
-        throw new RuntimeException(
-            "Could not find generic type for property: "
-                + propertyName
-                + " in class: "
-                + mainObj.getClass().getSimpleName());
-      }
-
-    } catch (Exception e) {
-      throw new RuntimeException(e);
-    }
-  }
-
-  private Class<?> getPropertyClass(Object obj, String propertyName) {
-    Assert.notNull(obj, "obj must not be null");
-    Assert.notNull(propertyName, "propertyName must not be null");
-
-    List<PropertyInfo> propertyInfoList = getObjectPropertyInfo(obj);
-    PropertyInfo propertyInfo =
-        propertyInfoList
-            .stream()
-            .filter(pi -> propertyName.equals(pi.getPropertyName()))
-            .findAny()
-            .orElse(null);
-
-    if (propertyInfo == null) {
-      throw new RuntimeException(
-          "property:" + propertyName + " not found in class:" + obj.getClass().getName());
-    } else {
-      return propertyInfo.getPropertyType();
-    }
   }
 }
